Retry weak weetbix exonerations regardless of what findit determines
When findit exonerates all failling tests, no retries will be
triggered. The recipe test that was supposed to cover this was
running because of an "invalid" mock test that forced it to always
trigger. Also rename the logging step since we probably shouldn't
trigger retries when the builder is explicitly setup to not allow
retries (and this isn't checked at the logging step)

Bug: 1314194
Change-Id: I014028ce923c2dd8fac21b6c008232ae841add80
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3857484
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Struan Shrimpton <sshrimp@google.com>