Update expectation_file_scripts gclient config with the bootstrapper.
For a recipe to be bootstrappable, it must update the gclient config
with the revision information provided by the bootstrapper so that the
properties and the checked out source are from the same revision.
Otherwise it's possible that the recipe will use a revision that landed
after the bootstrapper read the properties where the revision is
incompatible with the properties.

Change-Id: I9d6d80de672660c0e20094c18d47345e81b5ce27
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3864359
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>