[flakiness] Remove |current_tests| set
This change makes flakiness module compare and add new tests to
|preliminary_new_tests| on the fly. This saves memory and makes
it possible to only note down statistics of new tests. We will use
the statistics to make a decision for tests that's already flaky in
"with patch" and "retry shards with patch" test steps.


Bug:1291610
Change-Id: Ide86b428819b8b1e672041bc7b23d988e37770b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3867355
Reviewed-by: Jeff Yoon <jeffyoon@google.com>
Commit-Queue: Zhaoyang Li <zhaoyangli@chromium.org>