[v8] Add support to specify the source of a new version
This CL adds support for defining the version source in the recipe properties.

Additionally, instead of defining the name of the dependency to be trusted, we define the repository source instead. The dependency name might differ across multiple projects, while the url points to the project for which we need to decide if it can be trusted.

Bug: 1349706
Change-Id: Id1f7b1c5bf7f6d693925255f4cc8e1b2a98775e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3869148
Reviewed-by: Liviu Rau <liviurau@google.com>
Commit-Queue: Alexander Schulze <alexschulze@chromium.org>