Remove unnecessary venv argument from presubmit call.
The presubmit function accepts arbitrary kwargs that currently get
passed on to a python call and then to a step call with venv removed.
However, before the python call, venv is forced to True, so setting it
in the call has no effect and prevents switching the presubmit call to
use the step API where it is currently using the python API. Removing
the unnecessary venv argument will allow for the presubmit code to
remove its dependency on the python module.

Change-Id: I27540d867d76bf3b039a060ec98c6d89f462fe7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3887321
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@google.com>