Remove annotation_utils.py.
The annotation_utils module provided the functionality for creating
steps that was used with buildbot/kitchen. Fully migrated to
LUCI/bbagent, the annotations are not used. The only place where
annotation_utils was used was in runtest.py. A new function
_report_outcome was added to runtest.py to replace
annotation_utils.annotate. It prints out the same information without
the annotation characters.

Bug: 1008256
Change-Id: Iff8f4879c9100d4f4f4dbe35c1153f6b09ea368f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3887739
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>