Remove 'all_tests' field in the serialized_results log for all tests
The '$debug - all results' step dumps the recipe's internal
representation of the build's test results to a log. This can be
useful when debugging. But the presence of the 'all_tests' field
inflates the size of the log and makes it difficult to parse. eg:
https://logs.chromium.org/logs/chromium/buildbucket/cr-buildbucket/8803193553556709361/+/u/l_debug_-_all_results__with_patch_/serialized_results

Presumably listing all test names in a suite isn't very useful. So
this removes it for readability's sake.

Recipe-Nontrivial-Roll: build_limited
Recipe-Nontrivial-Roll: chrome_release
Bug: 1359508
Change-Id: I2e5ea2828d1529de0570656507a32ee9c6e55c37
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3893309
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>