Remove env.py and test_env.py.
env.py once existed to allow downstream repos to add additional
directories to the python path when using the depot_tools fetch, but
builders that check out the recipe repos no longer use fetch, so it is
only used via the test_env import in unittests and invoked via a
PRESUBMIT script to determine the paths to pass to pylint. This change
modifies the tests to do direct path manipulation and the PRESUBMIT
script to explicitly set the additional paths.

Bug: 1008256
Change-Id: I029592ca63fee56535f5d563f525857c06c25a0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3891965
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>