Set the revision for a repo to the commit ref if ID is not present.
In the case of a bootstrapped try builder for a dependency project
being run on a CL for the dependency repo, the input commit will be a
commit for the top level repo with only the ref set. In order to
ensure that the correct ref is checked out for the top-level repo, the
revision should be set to the ref rather than an empty ID value.

Bug: 1367346
Change-Id: I724c86d445963ee54879f3fb312d80a2c4143664
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3916692
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>