[fuchsia] Only fails fuchsia-binary-size on Failure of diff step.
Previously, this was a Warning step. However, these are apparently
treated as failures overall, which is not the intention of this (it
should be a success/the trybot should pass). This will not prevent
fuchsia-binary-size bot from failing unless the step is marked as
failure.

This should not affect the android-binary-size bot.

Change-Id: I9d139727f4e655615512967b05335404d53c49cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3919672
Reviewed-by: Stephanie Kim <kimstephanie@google.com>
Reviewed-by: Zijie He <zijiehe@google.com>
Commit-Queue: Rohan Pavone <rohpavone@chromium.org>