Fix builder links when reporting builders for standalone try builder.
The report builders code assumes the convention that any mirrored
builders for a try builder will be in the ci bucket. In the case of a
standalone try builder (one where the mirror is the same builder), this
won't be the case, the link should be back to the same builder. This
changes it so that report_builders takes the builder ID and when
creating the link for a mirrored builder, if it matches the builder ID
of the running builder, then it won't transform the bucket.

Change-Id: I4b9d002f1f54e66752594147127e3af98acfaec4
Recipe-Manual-Change: build_limited
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3924219
Reviewed-by: Gregory Guterman <guterman@google.com>
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>