[V8] Implement flake progression testing
Adds a new bisector type that for progression testing that checks
after calibration if a flake is fixed at ToT. If fixed, it bisects
into the range of known-bad...ToT and determines the commit that
fixed the flake.

This CL reuses existing bisection helpers for exercising the
bisection. Most new code is for fetching revisions from ToT and
calculating the relative offset to the known bad revision. The
code does not rely on commit positions, but instead fetches all
revisions between ToT and the known bad revision.

Internally all revisions are represented by offsets to the known
bad revision at offset 0. Previously with only regression bisection,
positive offsets referred to older commits. We kept this convention
and negative offsets refer to newer commits now, for the range
between known bad and ToT. We might want to turn this around in
the future in a separate CL as it is slightly unintuitive.

Bug: v8:13332
Change-Id: I3c85f0ec7ed4a4113d8cb531f2f80194b329f0c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3922134
Reviewed-by: Alexander Schulze <alexschulze@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>