[V8] Flako - introduce bisect modes
This deprecates the boolean properties for different modes and instead
introduces a property "mode" for regression, progression and repro
mode.

The repro case is aligned with existing bisector classes with a slight
logical change: It will now always print the flakes.pyl output step
as this simplifies the code and doesn't hurt.

Bug: v8:13332
Change-Id: I89f747e542b5b38a4a931a5687adb84572bc71ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3931546
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Liviu Rau <liviurau@google.com>