[V8] Flako - add optional initialization to ToT
This adds an on-demand initialization of the starting revision if
it is not given, which eases reproduction testing on ToT.

This also prepares renaming the to_revision property to just
revision, as the other is confusing with progression testing.

The flake command line and the flake verifier are already updated,
but we keep support for the old property for a grace period, so
that current flako command lines keep working.

Bug: v8:13332
Change-Id: Ia6185e6f5796ca2870f91f28ca8848384478b34a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3931551
Reviewed-by: Liviu Rau <liviurau@google.com>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>