[Flaky Reproducer] Remove bad symlink when repakcing test binary.
Chromite includes a symlink which points to a file it expects to exist
in a chroot. We aren't using chromite in a chroot, so this is an invalid
symlink. This causes `cas archive` commands which have this directory to
fail, so for now we're removing this directory if we see it in any
isolate we modify.
See https://crbug.com/1298283 for more detail.

Bug: 1366931
Change-Id: I0ad2b4031b85a52814713168b5877fd3c3d854fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3938482
Reviewed-by: Zhaoyang Li <zhaoyangli@chromium.org>
Commit-Queue: Kuan Huang <kuanhuang@chromium.org>