[v8] Add checkout_fuchsia_no_hooks in chromium checkout
The fuchsia dependency was removed from chromium/src in [1]. Our rollers do not update fuchsia anymore after this CL.

More changes have been made to chromium/src/DEPS in [2,3] to fix the skia roller. A new condition `checkout_fuchsia_no_hooks` got introduced.

Before [1], the condition `checkout_fuchsia` was fulfilled, since we use `gclient revinfo --deps all` [4] to retrieve deps, which sets `checkout_fuchsia` [5].

We cannot adapt the same approach to set checkout_fuchsia_no_hooks, as the name is not referring to a platform in gclient. Therefore, this CL sets a custom variable to the gclient solution when retrieving dependencies.

Note: We set this variable for all roller instances using chromium/src, though it is only needed for v8 (for now). However, the change is not checking out fuchsia, but only used by gclient to determine available dependencies from the DEPS file.

A LED run with this change was done in [6] and rolls fuchsia as part of the trusted rolls again. The CL cannot be submitted, and the build fails at this point, as this requires extended permissions.

[1] https://crrev.com/c/3914609
[2] https://crrev.com/c/3918832
[3] https://crrev.com/c/3924325
[4] https://crsrc.org/b/recipes/recipes/v8/auto_roll_incoming_deps.py;l=224;drc=567cea7c
[5] https://crsrc.org/d/gclient.py;l=1351;drc=1b8211ff
[6] https://chrome-swarming.appspot.com/task?id=5de4cc9fbd8e4b10

Bug: 1373807
Change-Id: Ic93145f421b5be95ac98e64ed0eb27a9901a3bab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3948864
Auto-Submit: Alexander Schulze <alexschulze@chromium.org>
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>