[resultdb] Add ResultDB support under a flag
ResultsDB support is not fully and neatly implemented and this is why we add it under a flag. We start by enabling ResultDB support in tester task execution for CQ bots. The feature is controlled with 'resultdb' flag activated with V8-Recipe-Flags footer in Gerrit CLs.

To led test this we launched a try bot in patch 10 of CL #1. Then launched a led test #2 from that bb id #3. That build triggers a tester run #4. The tester run triggers swarming tasks (ex. #5) that run with rdb wrapper around the test runner. I picked up the invocation name #6 from the logs of #4. Test results can be seen in Milo at #7. As this is still work in progress we are yet to have all results aggregated in Milo. So far we can only see a mysterious subset of what should actually be there.


#1 https://chromium-review.googlesource.com/c/v8/v8/+/3905385/35
#2 https://chromium-swarm.appspot.com/task?id=5de8d659cf531510
#3 8800495917994334849
#4 https://chromium-swarm.appspot.com/task?id=5de8d98ebce24610
#5 https://chromium-swarm.appspot.com/task?id=5de8da78c4e54510
#6 invocations/task-chromium-swarm.appspot.com-5de8d98ebce24611
#7 https://luci-milo.appspot.com/ui/inv/task-chromium-swarm.appspot.com-5de8d98ebce24611/test-results

Recipe-Nontrivial-Roll: build_limited
Bug: v8:13316
Change-Id: Ia15bc2424d00dc1bab9c71341af84a7364942643
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3925576
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Commit-Queue: Liviu Rau <liviurau@google.com>