Remove archive_utils.py and archive_utils_unittest.py.
archive_utils was unused except for in archive_utils_unittest.py.
archive_utils_unittest is not run as part of the presubmit and is only
imported by archive_build_unittest.py to access a couple of helper
functions (one of which was unnecessary). The helper function that was
necessary was moved to archive_build_unittest.py.

Bug: 1008256
Change-Id: I432462f36d007064172d309ab3c30c4b4cbee4ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3965368
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>