[V8] Restore ownership of V8 recipe module
After https://crrev.com/c/3956740 the V8 team lost the ownership of
the V8 module. This restores the owners file of V8 for now and
adds machenbach@ to the file ownership of v8*. We're going to test
if the owners plugin has a bug or if we need to update the configs.

Bug: 890222
Change-Id: I3468f5f8e8fd69f74e0eab84abe65325acca6ad0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3965856
Auto-Submit: Michael Achenbach <machenbach@chromium.org>
Commit-Queue: Takuto Ikuta <tikuta@chromium.org>
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>