[V8] Separate v8 and v8_tests recipe modules
This extracts testing related logic from the V8 recipe module into
its own module. This will make it easier to drop some historical
baggage in the new module v8_tests and to decouple it as much as
possible from the v8 recipe module.

This also prepares orchestrator/compilator, where the orchestrator
will only need to depend on v8_tests.

This is intended as a refactoring with no logical changes, though
the CL drops a few lines of dead code from obsolete features.

We keep convenience methods to keep the API compatible with
downstream dependencies. We remove them in a follow up once
downstream dependencies are updated.

This CL does not intend to improve the code yet, but merely moves
it and makes it work again. Improvements will be added in follow
up CLs.

Review instructions:
Patch 1 is a 1:1 move of connected blocks of code from v8 to v8_tests.
No changes to code applied.
Patch 2..last provides the modifications to review.

Bug: 890222
Change-Id: I5dd0a40a38afb88e002e0642d4ff6ee1e0ea7efe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3959884
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Alexander Schulze <alexschulze@chromium.org>