Add a no-op step to all builds right after 'Tests statistics'
We're seeing inflated step times for 'Tests statistics'. eg:
https://ci.chromium.org/ui/p/chromium/builders/try/win10_chromium_x64_rel_ng/1348560/overview

This adds a no-op step right after it to determine if the fault
actually lies with the 'Tests statistics' step, or if it's something
that occurs after that step completes. If that's the case, we should
see the time inflation shift from 'Tests statistics' to this new step.

Recipe-Nontrivial-Roll: build_limited
Bug: 1379170
Change-Id: Ibab35c0c9c0c37eb286962030975162b46192d2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3988682
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Stephanie Kim <kimstephanie@google.com>