[V8] Implement combined bisect mode for flako
This adds a mode "combined" that enables running progression and
regression testing in one run.

We first run progression testing to check if the flake still
reproduces. If it doesn't we calculate the fixed range and quit.

If it does, we continue with usual regression testing and return
a combined report.

Internally, we now return recipe result objects from methods on
the critical path, instead of raising exceptions for normal
failure cases. This enables combining the result.

Bug: v8:13332
Change-Id: Ie23d63d86b14bb408e2d34f2593ba8dcf5be6305
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3981714
Reviewed-by: Alexander Schulze <alexschulze@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Liviu Rau <liviurau@google.com>