Chmod and clobber the cros_cache dir on builds with clobber config set
m108 (currently beta) and m109 (currently canary) are using a broken
CrOS SDK that will add a read-only directory. This makes subsequent
builds break when they try managing the contents of the SDK dir and/or
prune it, eg:
http://go/bbid/8798739871488360417

A fix has landed, but it will take time to roll to canary and will also
need to be merged into 108. In the meantime we can have the builds
heal themselves by manually chmod'ing the entire SDK dir then just
removing it. But this will only be done when the 'clobber' config is
set. Most bots do not set it, but all official bots do. Since it
appears to just be official that's failing, this should work well.

Led build with this CL (had to apply the 'clobber' config):
https://luci-milo.appspot.com/raw/build/logs.chromium.org/chromium/led/chromium-orchestrator_chops-service-accounts.iam.gserviceaccount.com/a67fd4bd83774d6c0750d666ef46c5233a79b26f723856610a8e2e3f4b664aef/+/build.proto

Recipe-Nontrivial-Roll: chrome_release
Recipe-Nontrivial-Roll: build_limited
Bug: b/256012263
Change-Id: I895173ce6079b7e7d3e84d4dacd177fa5c37ab35
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3997067
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>