Refactor filter module.
This encapsulates more of the logic into smaller methods that analyze
makes use of instead of having one large method that contains nearly all
of the logic and that analyze passes most of its own arguments to. This
also removes the unnecessary attributes and methods that are not
referred to by other modules.

Change-Id: I97b6b8d9dc3699882f0b3edf40811b8502c9d213
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3989896
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Stephanie Kim <kimstephanie@google.com>