Mark chromium_utils_test.py as a Py3-only test
https://crrev.com/c/3976148/5 theoretically made this unittest py3
compatible. But it's still running under py2 in presubmits. And when
it flakes, it appears to only flake under py2:
https://crrev.com/c/3995152?attempt=20&checksRunsSelected=Build%20Presubmit&tab=checks
(The `presubmit` vs `presubmit py3` steps don't control if the test is
invoked via py3. Have to click into the failure logs to see the
sys version.)

So this prevents it from running under py2 going forward in the hopes
that it avoids the flakes.

Bug: 1380461
Change-Id: I124acaf95d1127a53a62f808772f0c96538999e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/4000425
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>