Add a debug log to all builds that print mem usage of all test results
The test classes in test_utils/util.py are used to represent the
results of all tests in the chromium recipe. Each test result tracks
the names, statuses, durations, and various other information about
a single test case.

These classes used to be very minimal. But overtime more and more
fields are being added, especially to the bottom-most layer in
RDBPerIndividualTestResults. An object of this class is
initialized for every test case the recipe runs. With 1000s of test
cases (and likely much more) this has a significant impact on the mem
footprint of a build. So much so that prev builds used to encounter
OOM errors and fail, until their machines were provisioned with more
memory.

By logging the mem usage of the internal repr of all test results
in the recipe, we should get some insight into the relative size
of the objects. And by watching how the sizes grow, we can gauge the
impact of changes to the classes themselves.

Various led builds with varying amount of failed tests:
https://ci.chromium.org/swarming/task/5e15aaac00391a10?server=chromium-swarm.appspot.com
https://ci.chromium.org/swarming/task/5e15ad206a9b5110?server=chromium-swarm.appspot.com
https://ci.chromium.org/swarming/task/5e15b8eb3a72e410?server=chromium-swarm.appspot.com

Recipe-Nontrivial-Roll: build_limited
Recipe-Nontrivial-Roll: chrome_release
Bug: 1356042
Change-Id: Ie49d22adfe9baf0ad31135cad648488da5ca8b96
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3972320
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Stephanie Kim <kimstephanie@google.com>