[V8] Prepare renaming MB and test configs on the V8 side
This prepares naming builders in MB and test configs on the V8 side.
In the configs, they still use legacy names, which will be updated
by: https://crrev.com/c/4006615

For MB, this CL tweaks the config similar to the reclient feature.
We read the config and override the old with the new name. If the
new name is already present it's a no-op.

For builders.pyl we look for bot builder-name keys now (old and
new).

Both theses measures allow to rename the keys on the V8 side and
recipes are guaranteed to work pre/post renaming.

With this CL we are also finally able to move the builder-renaming
logic into the compilator recipe and simplify the v8 recipe module,
removing the last compilator-only logic from there.

Bug: chromium:890222
Change-Id: I8382767099aa240dfa98db599134a45b2977d345
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/4002978
Reviewed-by: Liviu Rau <liviurau@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>