Add a wrapper for buildozer.
Buildozer is primarily geared towards operating on bazel BUILD files. It
can be used for generic starlark files by sending the file content as
input to the command. Doing this however means that the file won't be
modified in place and the exit code is unusual because buildozer uses 0
exit code to indicate that everything worked fine and files were
modified. In the case of operating on stdin, files won't be modified and
if everything works fine an exit code of 3 will be used.

Since this is all rather error-prone, this script abstracts invoking
buildozer as well as tailoring the input format towards the use case of
builder config migrations. The wrapper takes as input a json file with a
structured definition of the edits to make builders. The wrapper uses a
crude heuristic to determine the file(s) that should be updated and then
makes multiple buildozer invocations to apply them.

Bug: 868153
Change-Id: I6fa182d623014148ae6281c9360394574d2cb0f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/4021715
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Auto-Submit: Garrett Beaty <gbeaty@google.com>