Remove validate_task_results() from steps.Test classes
This method appears largely useless. The only use it had was to store
an isolated script test's json output to pass it to blink's special
archive step.

But it can grab that json directly in the class's run() method. So
validate_task_results() shouldn't be needed, and removing it cleans
up the file a bit.

Bug: 1209328
Change-Id: Ia1ab06d7b807c8cd2bdf87f598a9756c3ec9f809
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/4025916
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>