Remove support for web URLs from extract_build.py.
extract_build.py is only being invoked to extract builds from GCS and
the code for handling web urls isn't correctly migrated to python3: the
urlretrieve function must be accessed via urllib.request instead of
directly on urllib.

Additionally, the code had a number of conditionals based on the value
of the failure variable, but the variable was initialized to False at
the start of each loop and never modified, so the conditionals would
either always be false or always be true. The code was simplified to
remove these conditionals.

Change-Id: Ib39f00feab324ee9828832b9fc1b65fb94201deb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/4029040
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>