Fix python versioning for sync_generated_files.
The codesearch module was switched to run its scripts with
python3/vpython3 in https://crrev.com/c/3931033, but the .vpython3 file
that was added set the python version to 2.7, which means it isn't
actually being executed with a python3 binary. Attempting to fix the
version in the .vpython3 file in https://crrev.com/c/4029411 caused
build failures.

This change sets the version in .vpython to be 2.7 and the version in
.vpython3 to be 3.8 and switches the recipe code to invoke
sync_generated_files.py using vpython so that the recipe code accurately
reflects reality.

Bug: 1366996
Change-Id: I7d73e0df7bfbcc3b199476919df8c7b9ac09a532
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/4031849
Reviewed-by: Josip Sokcevic <sokcevic@google.com>
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Josip Sokcevic <sokcevic@google.com>