Add json output to the builder config migration operation.
This updates the chromium_tests_builder_config_migration module so that
the migration operation can specify for the output to be in json format.
If json output is enabled, instead of producing text snippets that can
be copied and pasted into builder definitions, it will produce a json
file in the schema understood by buildozer_wrapper.py to automatically
update builder definitions.

Bug: 868153
Change-Id: If6c7bf0546aad533088b32afa77fba9ebfd9ba6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/4029037
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>