migrate results.html away from test-results.appspot.com
This is the first step of the migration, to make testing changes in
results.html easier.

In future results.html will be served directly at the link below, with a
json parameter and an invocation-id parameter.
https://chromium-layout-test-archives.storage.googleapis.com/results.html

An example url for full_results_jsonp.js:
https://chromium-layout-test-archives.storage.googleapis.com/linux-rel/1209926/blink_wpt_tests%20%28with%20patch%29/full_results_jsonp.js

layout-test-results.zip will not be archived once we prove everything
works.

Bug: 1382949
Change-Id: Icb04dbaba7932a3985851f8cec140d11f5e2853b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/4050198
Reviewed-by: Stephanie Kim <kimstephanie@google.com>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Weizhong Xia <weizhong@google.com>