Prevent the 'ci_only tests' step instructions from being cut-off in milo
See the 'ci_only tests' step in builds like:
https://ci.chromium.org/ui/p/chromium/builders/try/linux-chromeos-rel-compilator/402131/overview

The "Include-Ci-Only-Tests: true" text can be cutoff and hidden if the
window is sufficiently small.

So this inserts a new line prior to those instructions so that the
second sentence always shows up on a new line when the step is
expanded.

See https://ci.chromium.org/swarming/task/5eddbd30f603d710 with
this applied. Note though that build rendering for led builds is
different from normal builds. So we won't know for sure what this
will look like until it's committed.

Bug: 1368697
Change-Id: Ia44962892524c7c4838ba15e674fd0fbce7eaf2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/4062596
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Stephanie Kim <kimstephanie@google.com>