[V8] Implement generic 2-phase tasks and use them for gcmole
This adds a new test type executed in two phases, with a transition
phase in the middle:
- Trigger/collect tasks of phase 1.
- Perform local calculations.
- Trigger/collect tasks of phase 2.

This is instantiated with a new version of gcmole, which will
replace the predecessor after it's not used anymore on an active
V8 branch:
- Trigger/collect parallel calculation of partial callgraphs.
- Local calculations:
  - Download the relevant gcmole files.
  - Merge the callgraphs (V8-side script).
  - Upload new bundle again including the merged callgraph.
- Trigger/collect gcmole checking in parallel

GCMole testing is now also marked idempotent, since the analysis
is deterministic. This can save a lot of resources when triggering
runs on swarming with the same input files, as now existing tasks
will be reused under the hood.

For more details, refer to the design doc attached to the bug.

This also simplifies test groups and removes an outdated TODO.

Bug: v8:12660
Change-Id: Iacc6a0bbf3193bb693959c524a6ff4eead0b0f48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/4061311
Reviewed-by: Liviu Rau <liviurau@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>