[WebRTC] Build 'default' instead of 'all' on Android bots
Android unittests rely on '//base', so some targets that are never used
in WebRTC are derived from the gn analysis step.
Some of these targets (eg //base:base_javatests_lib) are currently
not compiling on WebRTC Android. But it's not worth the effort to fix
these to be compilable.
So, use a 'default' target when analyzing and building on Android.

Bug: webrtc:14743
Change-Id: Ie0ba6bcfd9d7631156a8aa2ba108f16800a466f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/4082128
Commit-Queue: Daniel.L (Byoungchan) Lee <daniel.l@hpcnt.com>
Reviewed-by: Jeremy Leconte <jleconte@google.com>