Add support for limiting polymorphic operations to a subset of testers.
Currently, if a tester is set as a target builder and the runner builder
attempts to perform compilation, then the operation will most likely
fail. This will happen because the builder config that gets looked up
will be for the parent builder, so it will use the parent builder's GN
args, but it won't have any necessary reclient/goma properties set.

The alternative would be to specify the parent builder as the target
builder. Then the necessary properties would be set for compilation, but
if the builder triggers multiple testers, then the targets for all of
those testers will be included. To enable performing the polymorphic
operations for just a subset of testers, this CL adds support for
filtering the testers of a target builder. If a target builder specifies
a tester filter, then the returned builder config will not include any
other testers in scope.

Bug: 1346396
Change-Id: Ibf09d364088a88d9a95137c707e77fd6012ebc3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/4086532
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Struan Shrimpton <sshrimp@google.com>
Reviewed-by: Struan Shrimpton <sshrimp@google.com>