Enable code_coverage for third_party/{dawn,angle,swiftshader}
Dawn, ANGLE, and SwiftShader are maintained and primarily developed by
Chromium, so for maintenance purposes can be treated as part of
Chromium, similar to third_party/blink.

Dawn coverage should allow us to see coverage data for the WebGPU tests
that are being added to the code-coverage CI bots, and the others are
useful for Dawn-Chromium integration and possible future tests.

Bug: chromium:1399354
Change-Id: If9ad500028c213db3506caac8f4f4919a7226dbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/4149468
Reviewed-by: Prakhar Asthana <pasthana@google.com>
Commit-Queue: Kai Ninomiya <kainino@chromium.org>