[presubmit] Add timeout as a property
In [1] we exceed the timeout for the presubmit builder. The project runs pytype within presubmit. While this is sufficiently fast after an initial run, the first run on a fresh bot exceeds the timeout of 8 minutes.

This change enables us to override the timeout in the builder config.

Additionally, we remove the deprecated `properties.tryserver` feature from the recipe:

* PS0..3: Adding the feature to pass the timeout property.
* PS3..HEAD: Removing deprecated `properties.tryserver` method (test code only).

[1] https://ci.chromium.org/ui/p/v8/builders/try/v8_presubmit/b8791733584921707585/overview

Bug: 1408048
Change-Id: I6823ca089fc0fb207aec08de151bcba9a64d33ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/4173627
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Auto-Submit: Alexander Schulze <alexschulze@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>