Add version control to results.html
We now have a new version of results.html. The new version
results.html is uploaded to a fixed path in gcs. We want to check
the version before upload and not accidentally upload the old
version to that location.

Fetch the version file from GCS first, and only upload a newer version
of results.html when we have a newer version.

Bug: 1406043, 1382949
Change-Id: Iaf6c451fc39478ea08bd17c9f3ba3839e1986bb3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/4201335
Commit-Queue: Weizhong Xia <weizhong@google.com>
Reviewed-by: Stephanie Kim <kimstephanie@google.com>