[js-code-coverage] Pass through src and out directory
Paths in javascript coverage are absolute, when passing them to Zoss
they need to be relative from the src directory. To achieve this pass
the source directory. Furthermore there are files which are dynamically
generated that reside in the out directory that need to excluded.
Unfortunately the out directory is a child of the src directory so we
need to know the out directory as well.

Bug: b:265968684
Test: https://ci.chromium.org/swarming/task/602f853e8b0f2310
Change-Id: Ia30a1eb97398cc8e100cf7b03ef6b7d985228218
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/4221117
Reviewed-by: Prakhar Asthana <pasthana@google.com>
Auto-Submit: Ben Reich <benreich@chromium.org>
Commit-Queue: Prakhar Asthana <pasthana@google.com>