[ts_mon] Use vpython3 if available
The script itself [1] needs to be migrated as well, but there is no point in using py2 if py2 is omitted on the bot. After this change, the python3-incompatibilities will be shown in a build instead of the missing vpython binary.

[1] https://crsrc.org/i/infra/tools/send_ts_mon_values/

Bug: 1411311
Change-Id: I4c3f10c162598a0dfb66f4f4f3c96e86fd0e2898
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/4224675
Auto-Submit: Alexander Schulze <alexschulze@chromium.org>
Commit-Queue: Alexander Schulze <alexschulze@chromium.org>
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>