Allow different deps cache dirs for reclient compare builders.
Currently the reclient v reclient compare recipe has a problem where the
deps cache is set to the same location for both builds in the recipe.
This doesn't work for a few reasons (second build gets dep cache hits it
can't access, and it should be shared in the first place as it makes the
performance comparison wrong).

This change lets the comparison recipe set a flag that indicates to the
reclient api that the deps cache directory should be appended with the
ninja-step-name for the build step.  Each build will have it's own build
cache after this.

Bug: b/267178049
Change-Id: I3f80b527871e133fc5c3ede6d05967c7e2bb8144
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/4207590
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Michael Savigny <msavigny@google.com>